requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
astroid==3.2.2
attrs==23.2.0
black==24.4.2
boto3==1.34.118
botocore==1.34.118
cachetools==5.3.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.6
deepgram-sdk==3.2.7
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-limiter==0.1.6
filelock==3.14.0
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
identify==2.5.36
idna==3.7
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.63
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.1
orjson==3.10.3
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.8.2
pre-commit==3.7.1
protobuf==5.27.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.2.2
pymongo==4.7.2
pytest==8.2.1
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
redis==5.0.4
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
schedule==1.2.2
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.3.0
tiktoken==0.7.0
tomlkit==0.12.5
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
verboselogs==1.7
virtualenv==20.26.2
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
